# docs/swagger.yaml
basePath: /
definitions:
  dto.ClubMessageResponse:
    properties:
      message:
        type: string
      timestamp:
        type: string
      user:
        $ref: '#/definitions/dto.User'
    type: object
  dto.ClubResponse:
    properties:
      created_at:
        type: string
      created_by:
        type: integer
      description:
        type: string
      id:
        type: integer
      is_private:
        type: boolean
      name:
        type: string
      number_of_members:
        type: integer
    type: object
  dto.CreateClubRequest:
    properties:
      description:
        type: string
      is_private:
        type: boolean
      name:
        type: string
    required:
    - name
    type: object
  dto.ErrorResponse:
    properties:
      error:
        example: could not parse data
        type: string
    type: object
  dto.LeaderboardEntryResponse:
    properties:
      current_streak:
        type: integer
      last_checkedin:
        type: string
      score:
        type: integer
      user:
        $ref: '#/definitions/dto.User'
    type: object
  dto.LoginRequest:
    properties:
      avatar_id:
        type: string
      password:
        example: strongpassword
        type: string
      username:
        example: john
        type: string
    required:
    - password
    - username
    type: object
  dto.LoginResponse:
    properties:
      message:
        example: login successful
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  dto.MemberResponse:
    properties:
      joined_at:
        type: string
      role:
        type: string
      user:
        $ref: '#/definitions/dto.User'
    type: object
  dto.MessageResponse:
    properties:
      message:
        example: user created successfully
        type: string
    type: object
  dto.SendMessageRequest:
    properties:
      message:
        type: string
    required:
    - message
    type: object
  dto.SignupRequest:
    properties:
      avatar_id:
        type: string
      password:
        example: strongpassword
        type: string
      username:
        example: john
        type: string
    required:
    - avatar_id
    - password
    - username
    type: object
  dto.User:
    properties:
      avatar_id:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
  dto.UserStatsDTO:
    properties:
      avatar_id:
        type: string
      current_streak:
        type: integer
      last_checkedin:
        type: string
      longest_streak:
        type: integer
      rank:
        type: integer
      score:
        type: integer
      user_id:
        type: integer
      username:
        type: string
    type: object
info:
  contact: {}
  description: API for KlubRanks leaderboard system
  title: KlubRanks API
  version: "1.0"
paths:
  /clubs:
    get:
      description: Get all clubs the user is a member of
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ClubResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user's clubs
      tags:
      - Clubs
    post:
      consumes:
      - application/json
      description: Create a club and add creator as admin
      parameters:
      - description: Create club payload
        in: body
        name: club
        required: true
        schema:
          $ref: '#/definitions/dto.CreateClubRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ClubResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new club
      tags:
      - Clubs
  /clubs/{clubId}/leaderboard:
    get:
      description: Fetch top N users sorted by score
      parameters:
      - description: Club ID
        in: path
        name: clubId
        required: true
        type: integer
      - default: 50
        description: Result limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.LeaderboardEntryResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get club leaderboard
      tags:
      - Leaderboard
  /clubs/{clubId}/leaderboard/score:
    post:
      description: Increment user's score
      parameters:
      - description: Club ID
        in: path
        name: clubId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update leaderboard score
      tags:
      - Leaderboard
  /clubs/{clubId}/members:
    get:
      parameters:
      - description: Club ID
        in: path
        name: clubId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.MemberResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get club members
      tags:
      - Clubs
    post:
      consumes:
      - application/json
      parameters:
      - description: Club ID
        in: path
        name: clubId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add member to club
      tags:
      - Clubs
  /clubs/{clubId}/messages:
    get:
      description: Fetch paginated messages for a club
      parameters:
      - description: Club ID
        in: path
        name: clubId
        required: true
        type: integer
      - default: 50
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ClubMessageResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get club messages
      tags:
      - Messages
    post:
      consumes:
      - application/json
      description: Send a chat message inside a club
      parameters:
      - description: Club ID
        in: path
        name: clubId
        required: true
        type: integer
      - description: Message payload
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/dto.SendMessageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ClubMessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Send message to club chat
      tags:
      - Messages
  /clubs/{clubId}/stats/{userId}:
    get:
      parameters:
      - description: Club ID
        in: path
        name: clubId
        required: true
        type: integer
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.UserStatsDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get club user stats with id
      tags:
      - Clubs
  /clubs/{clubId}/stats/me:
    get:
      parameters:
      - description: Club ID
        in: path
        name: clubId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.UserStatsDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get club user stats for current user
      tags:
      - Clubs
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT
      parameters:
      - description: User login payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Login user
      tags:
      - Auth
  /signup:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: User signup payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.SignupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new user
      tags:
      - Auth
securityDefinitions:
  BearerAuth:
    description: Type "Bearer {your JWT token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"